requirements.txt
pyyaml
paramiko
mysql-connector-python
undetected_chromedriver
beautifulsoup4